updated examples

+++ mainUbf1/main.cpp
@@ -16,15 +16,15 @@
 #include <cstdlib>
 
 // default background frame rate
-const int bgRate {10};
-mixr::simulation::Station* station {};
+const int bgRate{10};
+mixr::simulation::Station* station{};
 
 // station builder
 mixr::simulation::Station* builder(const std::string& filename)
 {
    // read configuration file
-   int num_errors {};
-   mixr::base::Object* obj {mixr::base::edl_parser(filename, factory, &num_errors)};
+   int num_errors{};
+   mixr::base::Object* obj{mixr::base::edl_parser(filename, factory, &num_errors)};
    if (num_errors > 0) {
       std::cerr << "File: " << filename << ", number of errors: " << num_errors << std::endl;
       std::exit(EXIT_FAILURE);
@@ -63,11 +63,11 @@ void updateDataCB(int msecs)
    glutTimerFunc(msecs, updateDataCB, msecs);
 
    // current time
-   const double time {mixr::base::getComputerTime()};
+   const double time{mixr::base::getComputerTime()};
 
    // compute delta time
-   static double time0 {time};   // N-1 Time
-   const double dt {static_cast<double>(time - time0)};
+   static double time0{time};   // N-1 Time
+   const double dt{time - time0};
    time0 = time;
 
    station->updateData(dt);
@@ -95,8 +95,8 @@ int main(int argc, char* argv[])
    station->event(mixr::base::Component::RESET_EVENT);
 
    // set timer for the background tasks
-   const double dt {1.0 / static_cast<double>(bgRate)};
-   const int msecs {static_cast<int>(dt * 1000)};
+   const double dt{1.0 / static_cast<double>(bgRate)};
+   const int msecs{static_cast<int>(dt * 1000)};
 
    // ensure everything is reset
    station->updateData(dt);
